Made it kinda work
needs 1 tick delay add cd class

+++ src/main/java/net/passerines/avians/constants/Stats.java
@@ -22,11 +22,11 @@ public class Stats {
     public final static Pair<NamespacedKey, PersistentDataType> MANA = new Pair<>(Util.getNamespacedKey("Mana"), PersistentDataType.FLOAT);
 
     public final static Pair<NamespacedKey, PersistentDataType> MANA_REGEN = new Pair<>(Util.getNamespacedKey("ManaRegen"), PersistentDataType.FLOAT);
-    public final static Pair<NamespacedKey, PersistentDataType> DEFENSE = new Pair<>(Util.getNamespacedKey("Defense"), PersistentDataType.FLOAT);
-    public final static Pair<NamespacedKey, PersistentDataType> STRENGTH = new Pair<>(Util.getNamespacedKey("Strength"), PersistentDataType.FLOAT);
-    public final static Pair<NamespacedKey, PersistentDataType> DEXTERITY = new Pair<>(Util.getNamespacedKey("Dexterity"), PersistentDataType.FLOAT);
+    public final static Pair<NamespacedKey, PersistentDataType> DEFENSE = new Pair<>(Util.getNamespacedKey("Defense"), PersistentDataType.INTEGER);
+    public final static Pair<NamespacedKey, PersistentDataType> STRENGTH = new Pair<>(Util.getNamespacedKey("Strength"), PersistentDataType.INTEGER);
+    public final static Pair<NamespacedKey, PersistentDataType> DEXTERITY = new Pair<>(Util.getNamespacedKey("Dexterity"), PersistentDataType.INTEGER);
     public final static Pair<NamespacedKey, PersistentDataType> SPEED = new Pair<>(Util.getNamespacedKey("Speed"), PersistentDataType.FLOAT);
-    public final static Pair<NamespacedKey, PersistentDataType> CRITDAMAGE = new Pair<>(Util.getNamespacedKey("CritDamage"), PersistentDataType.FLOAT);
+    public final static Pair<NamespacedKey, PersistentDataType> CRITDAMAGE = new Pair<>(Util.getNamespacedKey("CritDamage"), PersistentDataType.INTEGER);
     public final static Pair<NamespacedKey, PersistentDataType> CRITCHANCE = new Pair<>(Util.getNamespacedKey("CritChance"), PersistentDataType.FLOAT);
     public final static Pair<NamespacedKey, PersistentDataType> CRITEXECUTIONRATE = new Pair<>(Util.getNamespacedKey("CritExecutionRate"), PersistentDataType.FLOAT);
 }
